Move OpenSSL bindings to a separate package
