+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
-members = ["pkcs11", "cdylib", "tools/softhsm"]
-default-members = [".", "pkcs11", "cdylib"]
+members = ["pkcs11", "ossl", "cdylib", "tools/softhsm"]
+default-members = [".", "pkcs11", "ossl", "cdylib"]
 
 [workspace.package]
 version = "1.2.0"
@@ -9,6 +9,11 @@ description = "A PKCS #11 software token written in Rust"
 homepage = "https://github.com/latchset/kryoptic"
 repository = "https://github.com/latchset/kryoptic"
 license = "GPL-3.0-or-later"
+
+[workspace.dependencies]
+kryoptic = { version = "1.2.0", path = ".", default-features = false }
+ossl = { version = "1.2.0", path = "ossl", default-features = false }
+pkcs11 = { version = "1.2.0", path = "pkcs11", default-features = false }
 
 [package]
 name = "kryoptic"
@@ -29,18 +34,14 @@ test = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[build-dependencies]
-bindgen = "0.71"
-pkg-config = "0.3"
-
 [dependencies]
+ossl.workspace = true
+pkcs11.workspace = true
 asn1 = "0.21"
 bimap = "0.6.3"
 bitflags = "2.4.1"
-cfg-if = "1.0.0"
 constant_time_eq = "0.3.0"
 data-encoding = "2.4.0"
-getrandom = "0.3"
 hex = "0.4.3"
 itertools = "0.14.0"
 libc = "0.2.151"
@@ -49,7 +50,6 @@ num-bigint = "0.4.4"
 num-integer = "0.1.45"
 num-traits = "0.2.17"
 once_cell = "1.18.0"
-pkcs11 = { version = "1.2.0", path = "pkcs11" }
 rusqlite = { version = "0.31.0", optional = true }
 serde = { version = "1.0.180", features = ["derive"] }
 serde_json = "1.0.104"
@@ -57,14 +57,16 @@ serial_test = "3.1.1"
 simplelog = { version = "0.12.2", default-features = false, optional = true }
 toml = { version = "0.8.19", default-features = false, features = ["display", "parse"] }
 uuid = { version = "1.4.1", features = ["v4"] }
-vsprintf = { version = "2.0.0", optional = true }
+
+[dev-dependencies]
+ossl = { version = "1.2.0", path = "ossl", default-features = false, features = ["dummy-integrity"] }
 
 [features]
 aes = []
 ecc = []
 ecdsa = ["ecc"]
 ecdh = ["ecc"]
-eddsa = ["ecc"]
+eddsa = ["ecc", "ossl/ossl320"]
 ec_montgomery = ["ecc"]
 ffdh = []
 hash = []
@@ -90,18 +92,17 @@ kdf_all = ["hkdf", "pbkdf2", "sp800_108", "sshkdf", "tlskdf"]
 
 standard = ["sqlitedb", "ecc_all", "ffdh", "hash_all", "kdf_all", "rsa"]
 
-ecc_fips = ["ecdsa", "ecdh"]
-fips = ["sqlitedb", "rusqlite/bundled", "aes", "ecc_fips", "ffdh", "hash_all", "kdf_all", "rsa", "pqc"]
+fips = ["ossl/fips", "sqlitedb", "rusqlite/bundled", "aes", "ecc_all", "ffdh", "hash_all", "kdf_all", "rsa", "pqc"]
 
-dynamic = [] # Builds against system libcrypto.so
+dynamic = ["ossl/dynamic"] # Builds against system libcrypto.so
 
 pkcs11_3_2 = ["pkcs11/pkcs11_3_2"] # Exposes PKCS#11 3.2 Interfaces and new functions
 
-log = ["dep:log", "dep:simplelog", "dep:vsprintf"] #Enables error/info tracing
+log = ["dep:log", "dep:simplelog", "ossl/log"] # Enables error/info tracing
 slow = [] # Enables slow tests
 
-mlkem = ["pkcs11_3_2"]
-mldsa = ["pkcs11_3_2", "hash"]
+mlkem = ["pkcs11_3_2", "ossl/ossl350"]
+mldsa = ["pkcs11_3_2", "hash", "ossl/ossl350"]
 
 pqc = ["mlkem", "mldsa"]
