+++ ossl/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "ossl"
+
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[build-dependencies]
+bindgen = "0.71"
+pkg-config = "0.3"
+
+[dependencies]
+cfg-if = "1.0.0"
+getrandom = "0.3"
+libc = "0.2.151"
+log = { version = "0.4.27", default-features = false, features = ["std"], optional = true }
+once_cell = "1.18.0"
+vsprintf = { version = "2.0.0", optional = true }
+
+[features]
+ossl320 = [] # Requires at a minimum OpenSSL 3.2.0
+ossl350 = [] # Requires at a minimum OpenSSL 3.5.0
+dynamic = [] # Builds against system libcrypto.so
+fips = ["ossl350"] # Builds against sources and libfips.a instead of libcrypto
+log = ["dep:log", "dep:vsprintf"] # Error tracing using log crate
+dummy-integrity = [] # USE ONLY for testing as a dev-depenency
