+++ ossl/build.rs
@@ -1,40 +1,9 @@
-// Copyright 2023 Simo Sorce
+// Copyright 2025 Simo Sorce
 // See LICENSE.txt file for terms
 
 use std::env;
 use std::panic::set_hook;
 use std::path::{Path, PathBuf};
-
-struct Features {
-    fips: bool,
-    dynamic: bool,
-}
-
-impl Features {
-    fn to_bools() -> Features {
-        #[cfg(all(feature = "dynamic", feature = "fips"))]
-        compile_error!("features `dynamic` and `fips` are mutually exclusive");
-
-        #[cfg(all(
-            feature = "ecdh",
-            not(any(feature = "ecdsa", feature = "ec_montgomery"))
-        ))]
-        compile_error!(
-            "Feature 'ecdh' requires either 'ecdsa' or 'ec_montgomery'"
-        );
-
-        Features {
-            #[cfg(feature = "fips")]
-            fips: true,
-            #[cfg(not(feature = "fips"))]
-            fips: false,
-            #[cfg(feature = "dynamic")]
-            dynamic: true,
-            #[cfg(not(feature = "dynamic"))]
-            dynamic: false,
-        }
-    }
-}
 
 #[derive(Debug)]
 pub struct OsslCallbacks;
@@ -50,16 +19,12 @@ impl bindgen::callbacks::ParseCallbacks for OsslCallbacks {
     ) -> Option<bindgen::callbacks::IntKind> {
         if name == "OPENSSL_VERSION_NUMBER" {
             if value < OPENSSL_3_5_0 {
-                #[cfg(any(
-                    feature = "mlkem",
-                    feature = "mldsa",
-                    feature = "fips"
-                ))]
-                panic!("OpenSSL 3.5.0 or later is required for mlkem, mldsa or fips");
+                #[cfg(feature = "ossl350")]
+                panic!("OpenSSL 3.5.0 or later is required");
             }
             if value < OPENSSL_3_2_0 {
-                #[cfg(feature = "eddsa")]
-                panic!("OpenSSL 3.2.0 or later is required for eddsa");
+                #[cfg(feature = "ossl320")]
+                panic!("OpenSSL 3.2.0 or later is required");
             }
             if value < OPENSSL_3_0_7 {
                 panic!(
@@ -97,7 +62,7 @@ fn ossl_bindings(args: &[&str], out_file: &Path) {
         .expect("Couldn't write bindings!");
 }
 
-fn build_ossl(features: &Features, out_file: &Path) {
+fn build_ossl(out_file: &Path) {
     let sources = std::env::var("KRYOPTIC_OPENSSL_SOURCES")
         .expect("Env var KRYOPTIC_OPENSSL_SOURCES is not defined");
     let openssl_path = std::path::PathBuf::from(sources)
@@ -134,7 +99,7 @@ fn build_ossl(features: &Features, out_file: &Path) {
     let ar_path: std::path::PathBuf;
     let ar_name: &str;
 
-    if features.fips {
+    if cfg!(feature = "fips") {
         buildargs.push("enable-fips");
 
         defines.push_str(" -DOPENSSL_PEDANTIC_ZEROIZATION");
@@ -228,7 +193,7 @@ fn build_ossl(features: &Features, out_file: &Path) {
     );
 
     let mut args = vec![&include_path, "-std=c90"];
-    if features.fips {
+    if cfg!(feature = "fips") {
         args.push("-D_KRYOPTIC_FIPS_");
     }
 
@@ -267,16 +232,14 @@ fn set_pretty_panic() {
 fn main() {
     set_pretty_panic();
 
-    let features = Features::to_bools();
-
     let out_path = PathBuf::from(env::var("OUT_DIR").unwrap());
     let ossl_bindings = out_path.join("ossl_bindings.rs");
 
     /* OpenSSL Cryptography */
-    if features.dynamic {
+    if cfg!(feature = "dynamic") {
         use_system_ossl(&ossl_bindings);
     } else {
-        build_ossl(&features, &ossl_bindings);
+        build_ossl(&ossl_bindings);
     }
 
     println!("cargo:rerun-if-changed=build.rs");
