+++ src/fips/mod.rs
@@ -5,9 +5,9 @@ use crate::config::FipsBehavior;
 use crate::error::Result;
 use crate::mechanism::Mechanisms;
 use crate::object::{ObjectFactories, ObjectType};
-use crate::ossl::fips;
 use crate::token::Token;
 
+use ossl::fips;
 use pkcs11::*;
 
 pub(crate) mod indicators;
