+++ src/ossl/drbg.rs
@@ -8,15 +8,10 @@ use std::ffi::c_char;
 
 use crate::error::Result;
 use crate::mechanism::DRBG;
-use crate::ossl::bindings::*;
+use crate::ossl::common::*;
 
+use ossl::bindings::*;
 use pkcs11::*;
-
-#[cfg(not(feature = "fips"))]
-use crate::ossl::get_libctx;
-
-#[cfg(feature = "fips")]
-use crate::ossl::fips::*;
 
 /// Implements HMAC-DRBG using SHA-256 as the underlying hash function.
 #[derive(Debug)]
@@ -42,7 +37,7 @@ impl HmacSha256Drbg {
         unsafe {
             let rng_spec: &[u8; 10] = b"HMAC-DRBG\0";
             let rand = EVP_RAND_fetch(
-                get_libctx(),
+                osslctx().ptr(),
                 rng_spec.as_ptr() as *const c_char,
                 std::ptr::null(),
             );
@@ -180,7 +175,7 @@ impl HmacSha512Drbg {
         unsafe {
             let rng_spec = b"HMAC-DRBG\0";
             let rand = EVP_RAND_fetch(
-                get_libctx(),
+                osslctx().ptr(),
                 rng_spec.as_ptr() as *const c_char,
                 std::ptr::null(),
             );
