+++ src/ossl/hkdf.rs
@@ -15,11 +15,12 @@ use crate::hmac::hmac_size;
 use crate::mechanism::{Derive, MechOperation, Mechanisms};
 use crate::misc::*;
 use crate::object::{Object, ObjectFactories};
-use crate::ossl::bindings::*;
 use crate::ossl::common::*;
 #[cfg(feature = "fips")]
 use crate::ossl::fips::*;
 
+use ossl::bindings::*;
+use ossl::{EvpKdfCtx, OsslParam};
 use pkcs11::*;
 
 /// Represents an active HKDF operation state.
@@ -301,7 +302,8 @@ impl Derive for HKDFOperation {
         #[cfg(feature = "fips")]
         fips_approval_prep_check();
 
-        let mut kctx = EvpKdfCtx::new(name_as_char(OSSL_KDF_NAME_HKDF))?;
+        let mut kctx =
+            EvpKdfCtx::new(osslctx(), name_as_char(OSSL_KDF_NAME_HKDF))?;
         let mut dkm = vec![0u8; keysize];
         let res = unsafe {
             EVP_KDF_derive(
