+++ src/ossl/mod.rs
@@ -1,38 +1,10 @@
 // Copyright 2023 Simo Sorce
 // See LICENSE.txt file for terms
 
-/// This module is automatically generated by bindgen from the OpenSSL Headers
-/// and includes a selection of functions and other items needed to access the
-/// libcrypto/libfips function needed.
-pub mod bindings {
-    #![allow(non_upper_case_globals)]
-    #![allow(non_camel_case_types)]
-    #![allow(dead_code)]
-    #![allow(non_snake_case)]
-    include!(concat!(env!("OUT_DIR"), "/ossl_bindings.rs"));
-}
-
-use once_cell::sync::Lazy;
-
-/// A structure representing the main crypto library context
-struct OsslContext {
-    context: *mut bindings::OSSL_LIB_CTX,
-}
-
-unsafe impl Send for OsslContext {}
-unsafe impl Sync for OsslContext {}
-
 /// The static instance of the library context lazily created on first use
-static OSSL_CONTEXT: Lazy<OsslContext> = Lazy::new(|| unsafe {
-    OsslContext {
-        context: bindings::OSSL_LIB_CTX_new(),
-    }
-});
-
-/// Accessor to get hold of the library context
-pub fn get_libctx() -> *mut bindings::OSSL_LIB_CTX {
-    OSSL_CONTEXT.context
-}
+#[cfg(not(feature = "fips"))]
+static OSSL_CONTEXT: ::once_cell::sync::Lazy<::ossl::OsslContext> =
+    ::once_cell::sync::Lazy::new(|| ::ossl::OsslContext::new_lib_ctx());
 
 pub mod aes;
 pub mod common;
